+++ include/qphix/invcg.h
@@ -37,7 +37,7 @@ class InvCG : public AbstractSolver<FT,
  public:
   typedef typename Geometry<FT, veclen, soalen, compress12>::FourSpinorBlock Spinor;
 
-  static constexpr uint8_t num_flav = EvenOddLinearOperatorBase::num_flav;
+  static constexpr int num_flav = EvenOddLinearOperatorBase::num_flav;
 
   InvCG(EvenOddLinearOperatorBase &M_, int MaxIters_)
       : M(M_), geom(M_.getGeometry()), MaxIters(MaxIters_)
